saving before checking types on trsp

+++ src/allpairs/src/floydWarshall.c
@@ -63,7 +63,7 @@ process(
 
     PGR_DBG("Load data");
     pgr_edge_t *edges = NULL;
-    int64_t total_tuples = 0;
+    size_t total_tuples = 0;
     pgr_get_data_4_columns(edges_sql, &edges, &total_tuples);
 
     if (total_tuples == 0) {
@@ -100,8 +100,8 @@ PGDLLEXPORT Datum
 #endif
 floydWarshall(PG_FUNCTION_ARGS) {
     FuncCallContext     *funcctx;
-    size_t              call_cntr;
-    size_t               max_calls;
+    uint32_t              call_cntr;
+    uint32_t               max_calls;
     TupleDesc            tuple_desc;
 
     /**************************************************************************/
@@ -139,7 +139,7 @@ floydWarshall(PG_FUNCTION_ARGS) {
         /*                                                                   */
         /*********************************************************************/
 
-        funcctx->max_calls = result_count;
+        funcctx->max_calls = (uint32_t)result_count;
         funcctx->user_fctx = result_tuples;
         if (get_call_result_type(fcinfo, NULL, &tuple_desc) != TYPEFUNC_COMPOSITE)
             ereport(ERROR,
