+++ src/dijkstra/src/many_to_many_dijkstra.c
@@ -66,10 +66,8 @@ static
 void
 process(
         char* edges_sql,
-        int64_t *start_vidsArr,
-        size_t size_start_vidsArr,
-        int64_t *end_vidsArr,
-        size_t size_end_vidsArr,
+        int64_t *start_vidsArr, size_t size_start_vidsArr,
+        int64_t *end_vidsArr, size_t size_end_vidsArr,
         bool directed,
         bool only_cost,
         General_path_element_t **result_tuples,
@@ -78,7 +76,7 @@ process(
 
     PGR_DBG("Load data");
     pgr_edge_t *edges = NULL;
-    int64_t total_tuples = 0;
+    size_t total_tuples = 0;
     pgr_get_data_5_columns(edges_sql, &edges, &total_tuples);
 
     if (total_tuples == 0) {
@@ -121,8 +119,8 @@ PGDLLEXPORT Datum
 #endif
 many_to_many_dijkstra(PG_FUNCTION_ARGS) {
     FuncCallContext     *funcctx;
-    size_t              call_cntr;
-    size_t               max_calls;
+    uint32_t              call_cntr;
+    uint32_t               max_calls;
     TupleDesc            tuple_desc;
 
     /**************************************************************************/
@@ -176,7 +174,7 @@ many_to_many_dijkstra(PG_FUNCTION_ARGS) {
         /*                                                                   */
         /*********************************************************************/
 
-        funcctx->max_calls = result_count;
+        funcctx->max_calls = (uint32_t)result_count;
         funcctx->user_fctx = result_tuples;
         if (get_call_result_type(fcinfo, NULL, &tuple_desc)
                 != TYPEFUNC_COMPOSITE) {
