+++ src/driving_distance/src/withPoints_dd.c
@@ -83,7 +83,7 @@ process(
     pgr_SPI_connect();
 
     Point_on_edge_t *points = NULL;
-    int64_t total_points = 0;
+    size_t total_points = 0;
     pgr_get_points(points_sql, &points, &total_points);
 
     char *edges_of_points_query = NULL;
@@ -95,11 +95,11 @@ process(
 
 
     pgr_edge_t *edges_of_points = NULL;
-    int64_t total_edges_of_points = 0;
+    size_t total_edges_of_points = 0;
     pgr_get_data_5_columns(edges_of_points_query, &edges_of_points, &total_edges_of_points);
 
     pgr_edge_t *edges = NULL;
-    int64_t total_edges = 0;
+    size_t total_edges = 0;
     pgr_get_data_5_columns(edges_no_points_query, &edges, &total_edges);
 
     PGR_DBG("freeing allocated memory not used anymore");
@@ -159,8 +159,8 @@ PGDLLEXPORT Datum
 #endif
 withPoints_dd(PG_FUNCTION_ARGS) {
     FuncCallContext     *funcctx;
-    size_t              call_cntr;
-    size_t               max_calls;
+    uint32_t              call_cntr;
+    uint32_t               max_calls;
     TupleDesc            tuple_desc;
 
     /*******************************************************************************/
@@ -205,7 +205,7 @@ withPoints_dd(PG_FUNCTION_ARGS) {
         /*                                                                             */
         /*******************************************************************************/
 
-        funcctx->max_calls = result_count;
+        funcctx->max_calls = (uint32_t) result_count;
         funcctx->user_fctx = result_tuples;
         if (get_call_result_type(fcinfo, NULL, &tuple_desc) != TYPEFUNC_COMPOSITE)
             ereport(ERROR,
